fix: WalkMessage 컴포넌트 분리
